Split up "speedup" tests for `_geometric_intersection` and `_surface_intersection`.
